requirements.txt
nfl_data_py
pandas
matplotlib
pyarrow
requests
python-dateutil

